// libs/shared-ui/ui/src/lib/components/kp-channel-card/models/kp-channel-card.model.ts
export interface KpChannelCardModel {
  id: string;
  name: string;
  category: string;
  cover_image: string;
  language: string;
  is_active: boolean;
  is_owner: boolean;
  is_contributor: boolean;
  subscription_id: string;
  stats: {
    pulses_count: number;
    subscribers_count: number;
    rating: number;
  };
}